+++ CHANGELOG.md
@@ -40,6 +40,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Issues with CurseForge returning bad data getting cached and wrong messages shown to the user
 - Description not being set from the mrpack when importing [#1004]
 - Shaderpacks not recognised when editing mods [#954]
+- Nested scroll panes causing issues with scrolling [#866]
 
 ### Misc
 - Update gradle wrapper version from 8.2 to 8.12